backend/src/requirements.txt
asgiref==3.4.1
astroid==2.8.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
braintree==4.12.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
distlib==0.3.2
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pylint==2.11.1
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
wrapt==1.12.1
